#4
Put in support for generating public/private keys, for managing pkc12
keystores (and serializing them to base64 so we can shove them into
our main store) as well as for moving around X.509 certs and creating
trust stores out of them. This sets us up for SSL support both on the
client and server.