Upgraded android tools to 0.7 and gradle to 1.9 and put in various work arounds for 1.9 bugs due to IntelliJ moving to Android Studio 0.4.
Thali Device Hub now seems to be successfully running the Thali server
in the background and enforcing our broken keying strategy.

Did a bunch of refactoring to get the generic Thali client code put
into the utilities projects. We are re-using that code to build the UX
for the Thali Device Hub.