First real checkin of Chrome Native Host
This is a bunch of code to let you set up a web page that can use pouch
to talk via a Chrome Native Host who will then handle authentication and
talk to a Thali Device Hub.

This code is just meant for demonstration purposes. There are tons of
insecure hacks in it and it's so inefficient it makes me cry. It's just
for our hack fest next week to give people a sense of what is possible.