Mostly updating code to work with CouchBase refactoring
Did work to move code into JavaUtilities from AndroidUtilities now that
the objects we need are available in a pure Java (e.g. no Android
dependencies) thanks to CouchBase's refactoring work.

Also moved ChromeExtension to be a top level directory in DotNetUtilities
to make it clear that it is separate from the Chrome Host.