Just a backup of the new web bridge
We are changing web bridge so that it uses pouchDB 2.0 with support for
custom xhr. We now use a httpkey handler we can register with pouchDB
so that our requests can now use proper httpkey URLs. We are also moving
the registration hack (e.g. sticking your key into the principal database
directly) off to the edge so that in general we expect xhr requests to
have 'proper' httpkey URLs. This means changing the Chrome Host so that
it expects a proper httpkey URL except when dealing with magic provision
methods.