Bug fixes
ThaliDeviceHubService - The signature of the listener had changed so I needed
to update it.
ThaliCryptoUtilities - Just stuck in an catch to help me with debugging
ConfigureRequestObjects - I screwed up and put the mkdir behind an assert.
This worked fine in Java land but Android by default (even in debug) doesn't
honor asserts so I wasn't creating the directories needed for tests. Oops.