Created AndroidPouchDbSDK & Build Java XML HTTP Request Bridge
This project is a scafolding for running HTML based PouchDB/Thali
apps on Android KitKat.

Unfortunately the androidTest in that project doesn't work for
reasons I haven't figured out yet.

I also enhanced utilities with functionality equivalent to the chrome
bridge written in C#. This will enable a WebView in Android or Java
to relay XHR requests via Java so we can set up the mutual SSL auth
connection for Thali.