Provide infrastructure to discover TDH's HTTPKEY URLs & nuke XHRBridge
I updated the core listener to produce a class that has its HTTPKEY URLs.

Then I updated the relay to take this structure and return it when asked
for the magic path.

Then I updated the Java TDH to save this structure out to a file in a
well known location in preparation for updating the Seed code to read it.

I also create a generic JavaThaliListenerContext just to clean things up
a bit on Java contexts. This let me delete the ContextInUserHomeDirectory
class and vastly simplify CreateContextInTemp.

I also nuked all the XmlHTTPRequestBridge code since we are abandoning
it in favor of the relay.