Fix what we broke in the revert
We did a revert because while restructuring the Thali projects using
gradle sub-projects made working with the project in the IDE much easier
it also raised two bugs. The first bug (which I'm just putting here so
we don't forget) is that when gradle creats POMs it needs to know the
groupid and version. We set these manually in the uploadArchive command
but that doesn't apply to dependencies. The fix was to put in the global
level something like:
subprojects {
  group = 'com.msopentech.thali'
  version = '0.0.2'
}
So we could fix that. The next bug we couldn't fix. It's actually a bug
in IntelliJ that when debugging sub-modules via Gradle it screws up the
paths and so resources aren't available, only classes. This killed our
ability to run our code in debug. Unfortunately IntelliJ, when we looked
up the link, made it clear they aren't going to fix it in IntelliJ 13 and
suggested using 14. The problem is that 14 is an Early Access Preview
which means (according to IntelliJ's own warning) that it isn't even
necessarily beta quality and anyway they are using a newer version of the
Android code that isn't compatible with the version of the Android tools
we are using. Since we aren't going to risk updating Android tools right
now this left us with no choice but to revert.
See http://youtrack.jetbrains.com/issue/IDEA-120939#comment=27-777233 for
details on the bug.

Also, while I was in the area, I updated the sub-projects gradles to 1.12
since reverting to the old structure means we will be building individual
projects as well as the globalBuild and we want to keep our gradle versions
consistent.

Jason also figured out a fix for the problem we were having with the
copyTestFiles task not running when we ran test in JavaUtilities, we
needed to hook to compileTestJava.