Have the TDH tell the relay Tor's sock port & pass in host/port for relay
The relay needs to talk to tor to resolve onion addresses into their
public keys. I forgot to hook that up. This checkin fixes that.

I also pass in the host/port for the relay. This is in preparation for
eventually making these dynamic.

Resolve https://github.com/thaliproject/thali/issues/12