New Coding Guidelines and tools!
This is really a proposal for coding guidelines and enforcement tools. If you don't like the guidelines or tools selected then speak now or forever hold your peace. We will start to enforce the use of these tools when we set up our continuous integration. I would expect that to be sometime after story 0.0.