Roughed out BLE/Multi-Peer/Wi-Fi Sections
Got the basics for how to bind discovery to our three favorite transports. Still need to specify how to switch from beacons to TLS. I'm explicitly not worrying about the specifics of how to make BLE work for iOS. We'll deal with that when we get to the BLE Hub and Android/iOS interop.